rpm: Fix handling of SOURCE_DATE_EPOCH
Contemporary versions of Fedora automatically set SOURCE_DATE_EPOCH
based on the changelog entry date stamp. In scenarios where it already
is defined, we do not want to redefine it.

Additionally, when building the libvirt package in an Open Build Service
instance, the spec file is not present in %_specdir, but instead in %_sourcedir.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>
Message-Id: <20201026220820.60737-1-ngompa13@gmail.com>