rpm: remove with_bash_completion condition
The %meson macro sets "--auto-features=enabled", thus any feature in the
RPM which has a "with_XXX" condition, needs to explicitly pass a
"-DXXX=state" arg to %meson to override the auto features setting.

The with_bash_completion condition is always set to 1, so rather than
adding an arg to %meson, just remove the condition.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20201028123037.666295-2-berrange@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>