Add support for specifying max physical address size.
This patch introduces the:

    <maxphysaddr mode='passthrough'/>
    <maxphysaddr mode='emulate' bits='42'/>

sub element of /domain/cpu.

Purpose is being able to have a guest see the physical address size
exactly as the host does (if mode='passthrough' is used) or any size
the user wants it to see (if mode='emulate' is used and bits='' is
specified).

This can be useful if the VM needs to have a large amount of memory.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Message-Id: <160398693628.12832.10312981876243507879.stgit@toolbox-dario-user-dev>