lxc: Cleanup after failed startup
If starting an container fails, the virLXCProcessStop() is
called. But since vm->def->id is not set until libvirt_lxc is
spawned (the domain's ID is PID of that process),
virLXCProcessStop() returns early as virDomainObjIsActive()
returns false. But doing so leaves behind resources reserved for
the containers during the startup process. Most notably, hostdevs
are not re-attached to the host, the domain's transient XML is
not removed, etc.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <c88b182339ad348a8ab9e0fd0703cd72a0cbc5cc.1604669413.git.mprivozn@redhat.com>