domain_conf.c: align pSeries mem modules in virDomainMemoryDefPostParse()
Until now, we were aligning just ppc64 NVDIMM modules in post parse
time. Let's extend the treatment to all ppc64 memory models. This
is possible because the alignment restrictions for the architecture
are imposed by the platform specification, being hypervisor-agnostic.

This will remove the need for mem module alignment for pSeries in
the QEMU driver, simplifying the code a bit.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20201111220725.356297-4-danielhb413@gmail.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>