qemu_domain.c: align x86 memory module in post parse time
This is already being done for pSeries guests, so let's extend
the treatment for x86. The difference here is that the existing
x86 alignment in the QEMU driver might be restricted only to
QEMU, meaning that we can't put it in virDomainMemoryDefPostParse()
without infering that all hypervisors would handle it the same
way.

Instead, let's create a qemuDomainMemoryDefPostParse() that will
be called in qemuDomainDeviceDefPostParse() to do that. After this
change, def->nmems will always be aligned after parsing time.
qemuDomainAlignMemorySizes() doesn't need to align each memory
module before calculating 'hotplugmem'.

We'll also generated aligned values in the resulting XML as well.
Some xml2xml tests were changed to reflect this new behavior.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20201111220725.356297-6-danielhb413@gmail.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>