qemu_domain.c: move size check for mem modules after alignment
The loop in which 'hotplugmem' is calculated in qemuDomainAlignMemorySizes()
is throwing a max size overflow error. This is a reminiscent of
the time where this loop was doing the memory module alignment.

It doesn't hurt to leave the check here, but it makes more sense to
move it to the places where the alignments are being done.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20201111220725.356297-7-danielhb413@gmail.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>