qemu_hotplug.c: avoid aligning the mem obj in qemuDomainDetachPrepMemory()
The alignment of the mem object in qemuDomainDetachPrepMemory()
allows for the virDomainMemoryFindByDef() call to find the
right memory object for detach.

Turns out that, after recent changes, the alignment of the mem
object is already granted at this point. This function is used by
qemuDomainDetachDeviceLive(), which in turn is used in two
places with a memory device:

- qemuDomainDetachDeviceLiveAndConfig(), after a virDomainDeviceDefParse()
calll that will align the mem object via PostParse callbacks;

- qemuDomainDetachDeviceAliasLiveAndConfig(), after a call to
virDomainDefFindDevice() that will get the already aligned mem object
from the domain definition.

This means that the qemuDomainMemoryDeviceAlignSize() call at the
start of this function is now obsolete and can be removed.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20201111220725.356297-8-danielhb413@gmail.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>